CHROMIUM: cpuquiet: runnable_threads: lower the CPU1 wake up threshold
Lower the CPU1 wake up threshold to improve the performance. This is the
benefit for the case when only CPU0 online and is handling some threads
which may consume too much CPU0 power. And the gltich could happen due
to that. So lower the threshold could make the CPU1 been invoked in time
and share the loading.

BUG=chrome-os-partner:48222
TEST=web audio is smooth

Change-Id: Ifcc9ec7cdda33d74d75646567e78e56d53ae1422
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/318164
Commit-Ready: Yomyung Leem <yungleem@chromium.org>
Tested-by: Kary Jin <karyj@nvidia.com>
Reviewed-by: Kary Jin <karyj@nvidia.com>
Reviewed-by: Yomyung Leem <yungleem@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>

default_thresholds[1] value has changed from 18 to 14 via commit
8919a0630028c32cf18f3cbbef3fc553dca51467 earlier.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>